refactor(runtime): split misc language helpers (#8197)
